refactor: introduce src/app/ for cross-domain orchestration
